requirements.txt
langgraph==0.6.11
langchain-core>=0.2.0,<1.0.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic==2.9.2
typer==0.12.5
rich==13.9.4
